hide exprtk from API
- add "internal" header folder which is not installed
- MathPluginFormulaHelper, which needs exprtk, is defined in internal header
