+++ CMakeLists.txt
@@ -76,10 +76,13 @@ endif()
 
 foreach(subdir ${subdirs})
   aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR}/${subdir}/src sources)
-  include_directories(${CMAKE_CURRENT_SOURCE_DIR}/${subdir}/include)
-  file(GLOB headers "${CMAKE_CURRENT_SOURCE_DIR}/${subdir}/include/*.h")
+  include_directories(${CMAKE_CURRENT_SOURCE_DIR}/${subdir}/include
+      ${CMAKE_CURRENT_SOURCE_DIR}/${subdir}/include/internal)
+  file(GLOB_RECURSE headers "${CMAKE_CURRENT_SOURCE_DIR}/${subdir}/include/*")
   set(SOURCE ${SOURCE} ${sources} ${headers})
-  install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/${subdir}/include/ DESTINATION include/ChimeraTK)
+  install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/${subdir}/include/
+      DESTINATION include/ChimeraTK
+      PATTERN "internal" EXCLUDE)
 endforeach()
 
 include(CheckLibraryExists)
