+++ device_backends/LogicalNameMapping/include/LNMMathPlugin.h
@@ -6,7 +6,6 @@
 
 #include <boost/make_shared.hpp>
 
-#include <exprtk.hpp>
 #include <utility>
 
 namespace ChimeraTK::LNMBackend {
@@ -60,46 +59,6 @@ namespace ChimeraTK::LNMBackend {
    private:
     // store weak pointer because plugin lifetime should not extend MathPluginFormulaHelper lifetime
     boost::weak_ptr<MathPluginFormulaHelper> _h;
-  };
-
-  /********************************************************************************************************************/
-
-  struct MathPluginFormulaHelper {
-    std::string varName;
-    exprtk::expression<double> expression;
-    exprtk::symbol_table<double> symbols;
-    exprtk::rtl::vecops::package<double> vecOpsPkg;
-    std::unique_ptr<exprtk::vector_view<double>> valueView;
-    std::map<boost::shared_ptr<NDRegisterAccessor<double>>, std::unique_ptr<exprtk::vector_view<double>>> params;
-
-    boost::shared_ptr<LogicalNameMappingBackend> _backend;
-    boost::shared_ptr<NDRegisterAccessor<double>> _target;
-    // We assume plugin lives at least as long as MathPluginFormulaHelper
-    MathPlugin* _mp;
-
-    MathPluginFormulaHelper(MathPlugin* p, const boost::shared_ptr<LogicalNameMappingBackend>& backend);
-
-    void compileFormula(const std::string& formula,
-        const std::map<std::string, boost::shared_ptr<ChimeraTK::NDRegisterAccessor<double>>>& parameters,
-        size_t nElements);
-
-    template<typename T>
-    void computeResult(std::vector<double>& x, std::vector<T>& resultBuffer);
-
-    // This function updates result in target based on latest values of parameter accessors and lastMainValue
-    void updateResult(ChimeraTK::VersionNumber versionNumber);
-
-    // Checks that all parameters have been written since opening the device.
-    // Returns false as long as at least one parameter is still on the backend's _versionOnOpen.
-    // Only call this function when holding the _writeMutex. It updates the _allParametersWrittenAfterOpen
-    // variable which is protected by that mutex.
-    bool checkAllParametersWritten();
-
-    //  only used if _hasPushParameter == true
-    std::vector<double> _lastMainValue;
-    ChimeraTK::DataValidity _lastMainValidity;
-
-    std::map<std::string, boost::shared_ptr<NDRegisterAccessor<double>>> _accessorMap;
   };
 
 } // namespace ChimeraTK::LNMBackend
